+++ osdep/macosx_menubar.m
@@ -715,7 +715,7 @@
     [input setPlaceholderString:@"URL"];
     [alert setAccessoryView:input];
 
-    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 0.1 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
+    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t) (0.1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [input becomeFirstResponder];
     });
