+++ player/client.c
@@ -899,7 +899,7 @@ static bool conv_node_to_format(void *dst, mpv_format dst_fmt, mpv_node *src)
     }
     if (dst_fmt == MPV_FORMAT_INT64 && src->format == MPV_FORMAT_DOUBLE) {
         if (src->u.double_ >= INT64_MIN && src->u.double_ <= INT64_MAX) {
-            *(int64_t *)dst = src->u.double_;
+            *(int64_t *)dst = (int64_t) src->u.double_;
             return true;
         }
     }
